backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
passlib
python-jose
python-dotenv
python-multipart
PyJWT
bcrypt==4.0.1
groq
sentence_transformers
faiss-cpu
numpy